# api/docs/swagger.yaml
basePath: /
definitions:
  handlers.MeResponse:
    properties:
      role:
        type: string
      username:
        type: string
    type: object
...
  handlers.RefreshRequest:
    properties:
      refresh_token:
        type: string
      username:
        type: string
    type: object
  handlers.RefreshTokenInfo:
    properties:
      expires_at:
        type: string
      ip_address:
        type: string
      issued_at:
        type: string
      session_key:
        type: string
      user_agent:
        type: string
      username:
        type: string
    type: object
  handlers.RegisterAsAdminRequest:
    properties:
      password:
        type: string
      role:
        description: e.g., "user" or "admin"
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API for managing inventory products and stock movements.
  title: Inventory Tracker API
  version: "1.0"
paths:
  /admin/bans:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "500":
          description: Redis error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List all currently banned users or IPs
      tags:
      - admin
  /admin/bans/{id}:
    delete:
      parameters:
      - description: User or IP to unban
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Ban removed
        "404":
          description: Not found
          schema:
            type: string
        "500":
          description: Redis error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Remove a ban for user or IP
      tags:
      - admin
  /admin/bans/summary/send:
    post:
      responses:
        "202":
          description: Ban summary sent
          schema:
            type: string
        "404":
          description: No bans logged today
          schema:
            type: string
        "500":
          description: Redis error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Send today's ban summary immediately
      tags:
      - admin
  /admin/tokens:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.RefreshTokenInfo'
            type: array
        "403":
          description: Forbidden
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List active refresh tokens
      tags:
      - admin
  /admin/tokens/{username}:
    delete:
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      responses:
        "204":
          description: Token revoked
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Revoke a user’s refresh token
      tags:
      - admin
  /admin/users:
    post:
      consumes:
      - application/json
      parameters:
      - description: User to create with role
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterAsAdminRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "409":
          description: User exists
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create user with custom role
      tags:
      - admin
  /admin/users/{username}/tokens:
    delete:
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      responses:
        "204":
          description: All sessions revoked
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: No active sessions
          schema:
            type: string
        "500":
          description: Internal error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Revoke all sessions for a user
      tags:
      - admin
    get:
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.RefreshTokenInfo'
            type: array
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: No sessions found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List refresh tokens for a specific user
      tags:
      - admin
    post:
      parameters:
      - description: Username to impersonate
        in: path
        name: username
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Failed to generate token
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Issue token for user (impersonation)
      tags:
      - admin
  /admin/users/{username}/tokens/{sessionKey}:
    delete:
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: Session key (IP+UA hash)
        in: path
        name: sessionKey
        required: true
        type: string
      responses:
        "204":
          description: Session revoked
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: User or session not found
          schema:
            type: string
        "500":
          description: Internal error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Revoke a specific session for a user
      tags:
      - admin
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: username and password
        in: body
        name: credentials
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Authenticate user and return JWT token
      tags:
      - auth
  /logout:
    post:
      responses:
        "204":
          description: Logged out
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Session not found
          schema:
            type: string
        "500":
          description: Internal error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Logout (invalidate refresh token for this session)
      tags:
      - auth
  /logout/all:
    post:
      responses:
        "204":
          description: All sessions revoked
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: No active sessions
          schema:
            type: string
        "500":
          description: Internal error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Logout from all devices (invalidate all refresh tokens)
      tags:
      - auth
  /me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MeResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get current user info
      tags:
      - auth
...
  /refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Invalid token
          schema:
            type: string
      summary: Refresh access token
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      parameters:
      - description: username and password
        in: body
        name: credentials
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            type: string
        "409":
          description: User exists
          schema:
            type: string
      summary: Register new user and return JWT token
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"